Slow Day with CSS
As I am watching the NFL Opener as well as just not knowing whats going on, I have been using a web inspector to figure out how to style things. Trying to make a box with the second image and paragraph and whatnot using a div, but there has been some unexpected behavior.