Scrapped and Carousel
Sorry about my mini hiatus! Busy with schoolwork. This commit is me kinda unsatisfied with what I had before, so I took the liberty of deleting most of the content, and tweaked it a bit. The container is still weird for the Ollie Intro div, but that is something I just slapped together real quick. The biggest change is fooling around witha  carousel. I am still trying to work out how to do it effectively, so bear with me!